fixed bug with int32 field in audit_logs table

+++ src/models/audit_log.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import uuid
 from datetime import datetime
 
-from sqlalchemy import DateTime, ForeignKey, Integer, String, func
+from sqlalchemy import BigInteger, DateTime, ForeignKey, Integer, String, func
 from sqlalchemy.dialects.postgresql import JSONB, UUID
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -20,7 +20,8 @@ class AuditLog(Base):
     event: Mapped[str] = mapped_column(String(128), nullable=False, index=True)
 
     actor: Mapped[str | None] = mapped_column(String(16), nullable=True)
-    actor_id: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    # Telegram user ids exceed int32, so this must be BIGINT.
+    actor_id: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
 
     master_id: Mapped[int | None] = mapped_column(ForeignKey("masters.id", ondelete="SET NULL"), nullable=True)
     client_id: Mapped[int | None] = mapped_column(ForeignKey("clients.id", ondelete="SET NULL"), nullable=True)
